1-preprocessing-pyspark/requirements.txt
pyspark>=3.4.0
Pillow>=9.0.0
pandas>=1.5.0
pyarrow>=12.0.0
